requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
altair>=5.0.0
